+++ test/spec/features/context-pad/ContextPadSpec.js
@@ -10,6 +10,8 @@ import {
   inject
 } from 'test/TestHelper';
 
+import { assign } from 'min-dash';
+
 import {
   query as domQuery,
   queryAll as domQueryAll,
@@ -55,9 +57,9 @@ describe('features/context-pad', function() {
     beforeEach(bootstrapDiagram({ modules: [ contextPadModule, initPadModule ] }));
 
 
-    function Provider(entries) {
+    function Provider(entriesOrUpdater) {
       this.getContextPadEntries = function(element) {
-        return entries || {};
+        return entriesOrUpdater || {};
       };
     }
 
@@ -67,11 +69,10 @@ describe('features/context-pad', function() {
       // given
       var provider = new Provider();
 
-      // when
-      contextPad.registerProvider(provider);
-
       // then
-      expect(contextPad._providers).to.eql([ provider ]);
+      expect(function() {
+        contextPad.registerProvider(provider);
+      }).to.not.throw;
     }));
 
 
@@ -93,6 +94,114 @@ describe('features/context-pad', function() {
       // pass over providers
       expect(provider.getContextPadEntries).to.have.been.calledWith('FOO');
     }));
+
+
+    describe('with updater', function() {
+
+      it('should allow to add entries', inject(function(contextPad) {
+
+        // given
+        function updater(entries) {
+          return assign(entries, { entryB: {} });
+        }
+
+        var plainProvider = new Provider({ entryA: { action: function() {} } }),
+            updatingProvider = new Provider(updater);
+
+        contextPad.registerProvider(plainProvider);
+        contextPad.registerProvider(updatingProvider);
+
+        // when
+        var entries = contextPad.getEntries();
+
+        // then
+        expect(entries.entryA).to.exist;
+        expect(entries.entryB).to.exist;
+      }));
+
+
+      it('should allow to update entries', inject(function(contextPad) {
+
+        // given
+        function updater(entries) {
+          return assign(entries, { entryA: { alt: 'text' } });
+        }
+
+        var plainProvider = new Provider({ entryA: { action: function() {} } }),
+            updatingProvider = new Provider(updater);
+
+        contextPad.registerProvider(plainProvider);
+        contextPad.registerProvider(updatingProvider);
+
+        // when
+        var entries = contextPad.getEntries();
+
+        // then
+        expect(entries.entryA).to.exist;
+        expect(entries.entryA).to.have.property('alt');
+      }));
+
+
+      it('should allow to remove entries', inject(function(contextPad) {
+
+        // given
+        function updater(entries) {
+          return {};
+        }
+
+        var plainProvider = new Provider({ entryA: { action: function() {} } }),
+            updatingProvider = new Provider(updater);
+
+        contextPad.registerProvider(plainProvider);
+        contextPad.registerProvider(updatingProvider);
+
+        // when
+        var entries = contextPad.getEntries();
+
+        // then
+        expect(entries.entryA).to.not.exist;
+      }));
+
+
+      describe('ordering', function() {
+
+        function updater(entries) {
+          return {};
+        }
+
+        var plainProvider = new Provider({ entryA: { action: function() {} } }),
+            updatingProvider = new Provider(updater);
+
+
+        it('should call providers by registration order per default', inject(function(contextPad) {
+
+          // given
+          contextPad.registerProvider(plainProvider);
+          contextPad.registerProvider(updatingProvider);
+
+          // when
+          var entries = contextPad.getEntries();
+
+          // then
+          expect(entries.entryA).to.not.exist;
+        }));
+
+
+        it('should call providers by priority', inject(function(contextPad) {
+
+          // given
+          contextPad.registerProvider(plainProvider);
+          contextPad.registerProvider(1200, updatingProvider);
+
+          // when
+          var entries = contextPad.getEntries();
+
+          // then
+          expect(entries.entryA).to.exist;
+        }));
+
+      });
+    });
 
 
     describe('entry className', function() {